% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archive.R
\name{Archive}
\alias{Archive}
\title{Container for objective function evaluations}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\arguments{
\item{unnest}{\code{character()}
Unnested columns are stored in seperate columns instead of list-columns.}
}
\description{
Container around a data.table which stores all performed \link{Objective} function calls by the \link{Evaluator}.

Returns data.table which contains all performed \link{Objective} function
calls.
}
\section{Construction}{
\preformatted{archive = Archive$new(search_space, codomain, minimize)
}
\itemize{
\item \code{search_space} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
Search space that is logged into archive.
\item \code{codomain} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
Codomain of objective function that is logged into archive.
\item \code{minimize} :: named \code{logical}.
Should objective (component) function be minimized (or maximized)?
}
}

\section{Fields}{

\itemize{
\item \code{search_space} :: \link[paradox:ParamSet]{paradox::ParamSet} from construction\cr
\item \code{codomain} :: \link[paradox:ParamSet]{paradox::ParamSet} from construction\cr
\item \code{minimize} :: named \code{logical}; from construction
\item \code{data} :: \link[data.table:data.table]{data.table::data.table}\cr
Holds data of the archive.
\item \code{n_evals} :: \code{integer(1)}\cr
Number of evaluations stored in the container.
}
}

\section{Methods}{

\itemize{
\item \code{add_evals(dt)}\cr
\link[data.table:data.table]{data.table::data.table} -> \code{self}\cr
Adds function evaluations to the archive table.
}
}

