% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveRush.R
\name{ArchiveRush}
\alias{ArchiveRush}
\title{Logging object for objective function evaluations}
\description{
Container around a \link[data.table:data.table]{data.table::data.table} which stores all performed
function calls of the Objective.
}
\section{S3 Methods}{

\itemize{
\item \code{as.data.table(archive)}\cr
\link{Archive} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a tabular view of all performed function calls of the
Objective. The \code{x_domain} column is unnested to separate columns.
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specification of the search space for the \link{Optimizer}.}

\item{\code{codomain}}{(\link{Codomain})\cr
Codomain of objective function.}

\item{\code{start_time}}{(\link{POSIXct})\cr
Time stamp of when the optimization started.
The time is set by the \link{Optimizer}.}

\item{\code{rush}}{(\code{Rush})\cr
Rush controller for parallel optimization.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all evaluations.}

\item{\code{n_evals}}{(\code{integer(1)})\cr
Number of evaluations stored in the archive.}

\item{\code{cols_x}}{(\code{character()})\cr
Column names of search space parameters.}

\item{\code{cols_y}}{(\code{character()})\cr
Column names of codomain target parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArchiveRush-new}{\code{ArchiveRush$new()}}
\item \href{#method-ArchiveRush-best}{\code{ArchiveRush$best()}}
\item \href{#method-ArchiveRush-nds_selection}{\code{ArchiveRush$nds_selection()}}
\item \href{#method-ArchiveRush-format}{\code{ArchiveRush$format()}}
\item \href{#method-ArchiveRush-print}{\code{ArchiveRush$print()}}
\item \href{#method-ArchiveRush-clear}{\code{ArchiveRush$clear()}}
\item \href{#method-ArchiveRush-freeze}{\code{ArchiveRush$freeze()}}
\item \href{#method-ArchiveRush-clone}{\code{ArchiveRush$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$new(search_space, codomain, rush)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies the search space for the \link{Optimizer}. The \link[paradox:ParamSet]{paradox::ParamSet}
describes either a subset of the \code{domain} of the \link{Objective} or it describes
a set of parameters together with a \code{trafo} function that transforms values
from the search space to values of the domain. Depending on the context, this
value defaults to the domain of the objective.}

\item{\code{codomain}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies codomain of function.
Most importantly the tags of each output "Parameter" define whether it should
be minimized or maximized.  The default is to minimize each component.}

\item{\code{rush}}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-best"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-best}{}}}
\subsection{Method \code{best()}}{
Returns the best scoring evaluation(s).
For single-crit optimization, the solution that minimizes / maximizes the objective function.
For multi-crit optimization, the Pareto set / front.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$best(n_select = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.}

\item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-nds_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-nds_selection}{}}}
\subsection{Method \code{nds_selection()}}{
Calculate best points w.r.t. non dominated sorting with hypervolume contribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$nds_selection(n_select = 1, ref_point = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.}

\item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.}

\item{\code{ref_point}}{(\code{numeric()})\cr
Reference point for hypervolume.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-format"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-format}{}}}
\subsection{Method \code{format()}}{
Helper for print outputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-print"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$print()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-clear}{}}}
\subsection{Method \code{clear()}}{
Clear all evaluation results from archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-freeze"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-freeze}{}}}
\subsection{Method \code{freeze()}}{
Copy the data from rush to a local \code{\link[data.table:data.table]{data.table::data.table()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$freeze()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveRush-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveRush-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArchiveRush$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
