% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_terminator.R
\name{as_terminator}
\alias{as_terminator}
\alias{as_terminator.Terminator}
\alias{as_terminators}
\alias{as_terminators.default}
\alias{as_terminators.list}
\title{Convert to a Terminator}
\usage{
as_terminator(x, ...)

\method{as_terminator}{Terminator}(x, clone = FALSE, ...)

as_terminators(x, ...)

\method{as_terminators}{default}(x, ...)

\method{as_terminators}{list}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}
}
\description{
Convert object to a \link{Terminator} or a list of \link{Terminator}.
}
\seealso{
\link{Terminator}
}
