% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{assign_result_default}
\alias{assign_result_default}
\title{Default assign_result function}
\usage{
assign_result_default(inst)
}
\arguments{
\item{inst}{\link{OptimInstance}}
}
\description{
Used internally in the \link{Optimizer}.
It is the default way to determine the result by simply obtaining the best performing result from the archive.
}
\keyword{internal}
