% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{bbotk-package}
\alias{bbotk}
\alias{bbotk-package}
\title{bbotk: Black-Box Optimization Toolkit}
\description{
Features highly configurable search spaces via the paradox package and optimizes every user-defined objective function. The package includes several optimization algorithms e.g. Random Search, Iterated Racing, Bayesian Optimization (in mlr3mbo) and Hyperband (in mlr3hyperband). bbotk is the base package of mlr3tuning, mlr3fselect and miesmuschel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbotk.mlr-org.com}
  \item \url{https://github.com/mlr-org/bbotk}
  \item Report bugs at \url{https://github.com/mlr-org/bbotk/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Martin Binder \email{martin.binder@mail.com}
}

Other contributors:
\itemize{
  \item Olaf Mersmann \email{olafm@statistik.tu-dortmund.de} [contributor]
}

}
