% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{bbotk.backup}
\alias{bbotk.backup}
\title{Backup Archive Callback}
\description{
This \link{CallbackBatch} writes the \link{Archive} after each batch to disk.
}
\examples{
clbk("bbotk.backup", path = "backup.rds")
}
