% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{bbotk_conditions}
\alias{bbotk_conditions}
\alias{error_bbotk}
\alias{error_bbotk_terminated}
\title{Condition Classes for bbotk}
\usage{
error_bbotk(msg, ..., class = NULL, signal = TRUE, parent = NULL)

error_bbotk_terminated(msg, ..., class = NULL, signal = TRUE, parent = NULL)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Error message.}

\item{...}{(any)\cr
Passed to \code{\link[=sprintf]{sprintf()}}.}

\item{class}{(\code{character})\cr
Additional class(es).}

\item{signal}{(\code{logical(1)})\cr
If \code{FALSE}, the condition object is returned instead of being signaled.}

\item{parent}{(\code{condition})\cr
Parent condition.}
}
\description{
Condition classes for bbotk.
}
\section{Errors}{

\itemize{
\item \code{error_bbotk()} for the \code{Mlr3ErrorBbotk} class,
signalling a general bbotk error.
\item \code{error_bbotk_terminated()} for the \code{Mlr3ErrorBbotkTerminated} class,
signalling a termination error.
}
}

