% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbotk_reflections.R
\docType{data}
\name{bbotk_reflections}
\alias{bbotk_reflections}
\title{Reflections for bbotk}
\format{
\link{environment}.
}
\usage{
bbotk_reflections
}
\description{
Environment which stores various information to allow objects to examine and
introspect their structure and properties
(c.f. \href{https://en.wikipedia.org/wiki/Reflective_programming}{Reflections}).

This environment be modified by third-party packages.

The following objects are set by \CRANpkg{bbotk}:
\itemize{
\item \code{optimizer_properties} (\code{character()})\cr
Properties of an optimizer, e.g. "dependencies".
\item \code{objective_properties} (\code{character()})\cr
Properties of an objective, e.g. "noisy".
}
}
\examples{
ls.str(bbotk_reflections)
}
\keyword{internal}
