% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bbotk_worker_loop}
\alias{bbotk_worker_loop}
\title{Worker loop for Rush}
\usage{
bbotk_worker_loop(rush, objective, search_space)
}
\arguments{
\item{rush}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}

\item{objective}{(\link{Objective})\cr
Objective function.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies the search space for the \link{Optimizer}. The \link[paradox:ParamSet]{paradox::ParamSet}
describes either a subset of the \code{domain} of the \link{Objective} or it describes
a set of parameters together with a \code{trafo} function that transforms values
from the search space to values of the domain. Depending on the context, this
value defaults to the domain of the objective.}
}
\description{
Loop run on the workers.
Pops a task from the queue and evaluates it with the objective function.
Pushes the results back to the data base.
}
