% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{branin}
\alias{branin}
\alias{branin_wu}
\title{Branin Function}
\source{
Wu J, Toscano-Palmerin S, Frazier PI, Wilson AG (2019).
\dQuote{Practical Multi-fidelity Bayesian Optimization for Hyperparameter Tuning.}
1903.04703.
}
\usage{
branin(x1, x2, noise = 0)

branin_wu(x1, x2, fidelity)
}
\arguments{
\item{x1}{(\code{numeric()}).}

\item{x2}{(\code{numeric()}).}

\item{noise}{(\code{numeric()})\cr
Noise term that is added to the function value.
Default is \code{0}.}

\item{fidelity}{(\code{numeric()}).}
}
\value{
\code{numeric()}
}
\description{
Classic 2-D Branin function with noise \code{branin(x1, x2, noise)} and Branin function with fidelity parameter
\code{branin_wu(x1, x2, fidelity)}.
}
\examples{
branin(x1 = 12, x2 = 2, noise = 0.05)
branin_wu(x1 = 12, x2 = 2, fidelity = 1)
}
