% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackAsync.R
\name{CallbackAsync}
\alias{CallbackAsync}
\title{Create Asynchronous Optimization Callback}
\description{
Specialized \link[mlr3misc:Callback]{mlr3misc::Callback} for asynchronous optimization.
Callbacks allow to customize the behavior of processes in bbotk.
The \code{\link[=callback_async]{callback_async()}} function creates a \link{CallbackAsync}.
Predefined callbacks are stored in the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_callbacks} and can be retrieved with \code{\link[=clbk]{clbk()}}.
For more information on optimization callbacks see \code{\link[=callback_async]{callback_async()}}.
}
\seealso{
\code{\link[=callback_async]{callback_async()}}
}
\section{Super class}{
\code{\link[mlr3misc:Callback]{mlr3misc::Callback}} -> \code{CallbackAsync}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{on_optimization_begin}}{(\verb{function()})\cr
Stage called at the beginning of the optimization in the main process.
Called in \code{Optimizer$optimize()}.}

\item{\code{on_worker_begin}}{(\verb{function()})\cr
Stage called at the beginning of the optimization on the worker.
Called in the worker loop.}

\item{\code{on_optimizer_before_eval}}{(\verb{function()})\cr
Stage called after the optimizer proposes points.
Called in \code{OptimInstance$.eval_point()}.}

\item{\code{on_optimizer_after_eval}}{(\verb{function()})\cr
Stage called after points are evaluated.
Called in \code{OptimInstance$.eval_point()}.}

\item{\code{on_optimizer_queue_before_eval}}{(\verb{function()})\cr
Stage called after the optimizer proposes points.
Called in \code{OptimInstance$.eval_queue()}.}

\item{\code{on_optimizer_queue_after_eval}}{(\verb{function()})\cr
Stage called after points are evaluated.
Called in \code{OptimInstance$.eval_queue()}.}

\item{\code{on_worker_end}}{(\verb{function()})\cr
Stage called at the end of the optimization on the worker.
Called in the worker loop.}

\item{\code{on_result_begin}}{(\verb{function()})\cr
Stage called before the results are written.
Called in \code{OptimInstance$assign_result()}.}

\item{\code{on_result_end}}{(\verb{function()})\cr
Stage called after the results are written.
Called in \code{OptimInstance$assign_result()}.}

\item{\code{on_optimization_end}}{(\verb{function()})\cr
Stage called at the end of the optimization in the main process.
Called in \code{Optimizer$optimize()}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CallbackAsync-clone}{\code{CallbackAsync$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Callback" data-id="call"><a href='../../mlr3misc/html/Callback.html#method-Callback-call'><code>mlr3misc::Callback$call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Callback" data-id="format"><a href='../../mlr3misc/html/Callback.html#method-Callback-format'><code>mlr3misc::Callback$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Callback" data-id="help"><a href='../../mlr3misc/html/Callback.html#method-Callback-help'><code>mlr3misc::Callback$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Callback" data-id="initialize"><a href='../../mlr3misc/html/Callback.html#method-Callback-initialize'><code>mlr3misc::Callback$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Callback" data-id="print"><a href='../../mlr3misc/html/Callback.html#method-Callback-print'><code>mlr3misc::Callback$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallbackAsync-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CallbackAsync-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallbackAsync$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
