% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimInstance.R
\name{choose_search_space}
\alias{choose_search_space}
\title{Choose Search Space}
\usage{
choose_search_space(objective, search_space)
}
\arguments{
\item{objective}{(\link{Objective}) The objective function.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet} | \code{NULL}) Optional explicit search space.}
}
\value{
A \link[paradox:ParamSet]{paradox::ParamSet} to use as the search space.
}
\description{
Determines the search space from an objective's domain, handling TuneTokens.
Used internally by \link{OptimInstance} and \link{OptimInstanceAsync}.
}
