% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextBatch.R
\name{ContextBatch}
\alias{ContextBatch}
\title{Batch Optimization Context}
\description{
A \link{CallbackBatch} accesses and modifies data during the optimization via the \code{ContextBatch}.
See the section on active bindings for a list of modifiable objects.
See \code{\link[=callback_batch]{callback_batch()}} for a list of stages which that \code{ContextBatch}.
}
\section{Super class}{
\code{\link[mlr3misc:Context]{mlr3misc::Context}} -> \code{ContextBatch}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instance}}{(\link{OptimInstance}).}

\item{\code{optimizer}}{(\link{Optimizer}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{xdt}}{(\link[data.table:data.table]{data.table::data.table})\cr
The points of the latest batch.
Contains the values in the search space i.e. transformations are not yet applied.}

\item{\code{result}}{(\link[data.table:data.table]{data.table::data.table})\cr
The result of the optimization.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextBatch-new}{\code{ContextBatch$new()}}
\item \href{#method-ContextBatch-clone}{\code{ContextBatch$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="format"><a href='../../mlr3misc/html/Context.html#method-Context-format'><code>mlr3misc::Context$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="print"><a href='../../mlr3misc/html/Context.html#method-Context-print'><code>mlr3misc::Context$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContextBatch-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextBatch$new(inst, optimizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inst}}{(\link{OptimInstance}).}

\item{\code{optimizer}}{(\link{Optimizer}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
