% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextOptimization.R
\name{ContextOptimization}
\alias{ContextOptimization}
\title{Optimization Context}
\description{
This \link{Context} can access the
\itemize{
\item \code{instance} - \link{OptimInstance}
\item \code{optimizer} - \link{Optimizer}
\item \code{result} - Optimization result
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{on_optimization_begin
    on_optimizer_before_eval
        on_eval_after_design
        on_eval_after_benchmark
        on_eval_after_aggregation
    on_result_begin
    on_result_end
    on_optimization_after_eval
on_optimization_end
}\if{html}{\out{</div>}}
}
\section{Super class}{
\code{\link[mlr3misc:Context]{mlr3misc::Context}} -> \code{ContextOptimization}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instance}}{(\link{OptimInstance})\cr
Instance.}

\item{\code{optimizer}}{(\link{Optimizer})\cr
Optimizer.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{(\link[data.table:data.table]{data.table::data.table})\cr
Result.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextOptimization-new}{\code{ContextOptimization$new()}}
\item \href{#method-ContextOptimization-clone}{\code{ContextOptimization$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="format"><a href='../../mlr3misc/html/Context.html#method-Context-format'><code>mlr3misc::Context$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="print"><a href='../../mlr3misc/html/Context.html#method-Context-print'><code>mlr3misc::Context$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextOptimization-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContextOptimization-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextOptimization$new(instance, optimizer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{(\link{OptimInstance})\cr
Instance.}

\item{\code{optimizer}}{(\link{Optimizer})\cr
Optimizer.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextOptimization-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextOptimization-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextOptimization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
