% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{evaluate_queue_default}
\alias{evaluate_queue_default}
\title{Default Evaluation of the Queue}
\usage{
evaluate_queue_default(instance)
}
\arguments{
\item{instance}{\link{OptimInstanceAsync}.}
}
\description{
Used internally in \verb{$.optimize()} of \link{OptimizerAsync} classes to evaluate a queue of points e.g. in \link{OptimizerAsyncGridSearch}.
}
\keyword{internal}
