% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluator.R
\name{Evaluator}
\alias{Evaluator}
\title{Evaluator for logged and parralel evaluations}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Wraps an \link{Objective} function with extra services for convenient evaluation.
\itemize{
\item Automatic storing of results in an \link{Archive} after evaluation.
\item Automatic checking for termination. Evaluations of design points are performed in batches.
Before a batch is evaluated, the \link{Terminator} is queried for the remaining budget.
If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.
\item Parallel evaluation of a batch via \link[future.apply:future.lapply]{future.apply::future.lapply}.
\item Logging of evaluations on R console
\item Encapsulated evaluation via evaluate / callr with function \link[mlr3misc:encapsulate]{mlr3misc::encapsulate}
to guard against exceptions and even segfaults.
Note that behavior is define in \link{Objective} argument \code{encapsulate}.
}
}
\section{Construction}{
\preformatted{ev = Evaluator$new(objective, archive, terminator)
}
\itemize{
\item \code{objective} :: \link{Objective}.
\item \code{archive} :: \link{Archive}.
\item \code{terminator} :: \link{Terminator}.
}
}

\section{Fields}{

\itemize{
\item \code{objective} :: \link{Objective}; from construction.
\item \code{terminator} :: \link{Terminator}; from construction.
}
}

\section{Methods}{

\itemize{
\item \code{eval_batch(dt)}\cr
\code{\link[data.table:data.table]{data.table::data.table()}} -> named \code{list()}\cr
Evaluates all hyperparameter configurations in \code{dt} through resampling, where each configuration is a row, and columns are scalar parameters.
eturns a named list with the following elements:
\itemize{
\item \code{"batch_nr"}: Number of the new batch.
This number is calculated in an auto-increment fashion.
\item \code{"perf"}: A \code{\link[data.table:data.table]{data.table::data.table()}} of evaluated performances for each row of the \code{dt}.
Has the same number of rows as \code{dt}, and the same number of columns as length of \code{measures}.
Columns are named with measure-IDs. A cell entry is the (aggregated) performance of that configuration for that measure.
}

Before each batch-evaluation, the \link{Terminator} is checked, and if it is positive, an exception of class \code{terminated_error} is raised.
This function should be internally called by the tuner.
}
}

