% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Feature.R
\name{Feature}
\alias{Feature}
\title{Feature}
\description{
Feature class for QDO. Describes the feature part of QDO.
Mainly consists of a feature function (\link{Objective}), niches (\link{Niches}), and a surrogate (\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate} | \code{NULL}).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{character(1)}).}

\item{\code{feature_function}}{(\link{Objective}).}

\item{\code{niches}}{\link{(Niches)}.}

\item{\code{surrogate}}{(\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate}).}

\item{\code{model_feature_function}}{(\code{logical(1)}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Feature$new()}}
\item \href{#method-eval_dt}{\code{Feature$eval_dt()}}
\item \href{#method-clone}{\code{Feature$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$new(id, feature_function, niches, surrogate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}).}

\item{\code{feature_function}}{(\link{Objective}).}

\item{\code{niches}}{(\link{Niches}).}

\item{\code{surrogate}}{(\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_dt"></a>}}
\if{latex}{\out{\hypertarget{method-eval_dt}{}}}
\subsection{Method \code{eval_dt()}}{
Evaluates all input values in \code{xdt}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$eval_dt(xdt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{\link[data.table:data.table]{data.table::data.table}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.table}\cr
The column has to have the same name as the id of the acq_fun, because we
renamed the id of the codomain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Feature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
