% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\name{Instance}
\alias{Instance}
\title{Instance Class}
\arguments{
\item{task}{(\link[mlr3:Task]{mlr3::Task}).}

\item{learner}{(\link[mlr3:Learner]{mlr3::Learner}).}

\item{resampling}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr
Note that uninstantiated resamplings are instantiated during
construction so that all configurations are evaluated on the same data
splits.}

\item{measures}{(list of \link[mlr3:Measure]{mlr3::Measure}).}

\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet}).}

\item{terminator}{(\link{Terminator}).}

\item{bm_args}{(named \code{list()})\cr
Further arguments for \code{\link[mlr3:benchmark]{mlr3::benchmark()}}.}

\item{...}{(ignored).}
}
\description{
This is the abstract base class for instance objects.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Helper for print outputs.

Printer.

Performace evaluator.

Objective function.

Archive.

Best \link[mlr3:ResampleResult]{mlr3::ResampleResult}.

The optimization algorithm writes the best found list of settings and
estimated performance values here. For internal use.
}
\section{Fields}{

\describe{
\item{\code{task}}{(\link[mlr3:Task]{mlr3::Task}).}

\item{\code{learner}}{(\link[mlr3:Learner]{mlr3::Learner}).}

\item{\code{resampling}}{(\link[mlr3:Resampling]{mlr3::Resampling})\cr}

\item{\code{measures}}{(list of \link[mlr3:Measure]{mlr3::Measure}).}

\item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet}).}

\item{\code{terminator}}{(\link{Terminator}).}

\item{\code{bm_args}}{(named \code{list()})\cr
Further arguments for \code{\link[mlr3:benchmark]{mlr3::benchmark()}}.}

\item{\code{bmr}}{(\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult})\cr
A benchmark result, container object for all performed
\link[mlr3:ResampleResult]{mlr3::ResampleResult}s when evaluating hyperparameter configurations.}

\item{\code{start_time}}{(\code{POSIXct(1)})\cr
Time the optimization was started. This is set in the beginning of
optimization.}

\item{\code{n_evals}}{(\code{integer(1)})\cr
Number of configuration evaluations stored in the container.}

\item{\code{result}}{(named \code{list()})\cr
Result of the optimization, i.e., the optimal settings and its estimated
performance:}
}}

\concept{Instance}
