% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{local_search_control}
\alias{local_search_control}
\title{Local Search Control}
\usage{
local_search_control(
  minimize = TRUE,
  n_searches = 10L,
  n_steps = 5L,
  n_neighs = 10L,
  mut_sigma_init = 0.1,
  mut_sigma_factor = 1.2,
  mut_sigma_max = 0.5,
  stagnate_max = 10
)
}
\arguments{
\item{minimize}{(\code{logical(1)})\cr
Whether to minimize the objective.}

\item{n_searches}{(\code{integer(1)})\cr
Number of local searches.}

\item{n_steps}{(\code{integer(1)})\cr
Number of steps per local search.}

\item{n_neighs}{(\code{integer(1)})\cr
Number of neighbors per local search.}

\item{mut_sigma_init}{(\code{numeric(1)})\cr
Initial mutation sigma.}

\item{mut_sigma_factor}{(\code{numeric(1)})\cr
Factor by which to adapt the mutation sigma.}

\item{mut_sigma_max}{(\code{numeric(1)})\cr
Maximum mutation sigma.}

\item{stagnate_max}{(\code{integer(1)})\cr
Maximum number of no-improvement steps for a local search before it is randomly restarted.}
}
\value{
(\code{local_search_control})\cr
List with control params as S3 object.
}
\description{
Control parameters for local search optimizer, see \code{\link[=local_search]{local_search()}} for details.
}
