% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr_assertions}
\alias{mlr_assertions}
\alias{assert_terminator}
\alias{assert_set}
\title{Assertion for bbotk objects}
\usage{
assert_terminator(terminator)

assert_set(x, empty = TRUE, .var.name = vname(x))
}
\arguments{
\item{terminator}{(\link{Terminator})}

\item{x}{(\link{any})}

\item{empty}{(\link{logical(1)})}

\item{.var.name}{(\link{character(1)})}
}
\description{
Most assertion functions ensure the right class attrbiture, and optionally
additional properties.

If an assertion fails, an exception is raised. Otherwise, the input object is
returned invisibly.
}
\keyword{internal}
