% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatchHEBO.R
\name{mlr_optimizers_hebo}
\alias{mlr_optimizers_hebo}
\alias{OptimizerBatchHEBO}
\title{Heteroscedastic Evolutionary Bayesian Optimization (HEBO)}
\description{
\code{OptimizerBatchHEBO} class that implements optimization via the Python
HEBO framework. Calls HEBO via \CRANpkg{reticulate}.
}
\section{Dictionary}{

This \link{Optimizer} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_optimizers} or with the associated sugar function \code{\link[=opt]{opt()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_optimizers$get("hebo")
opt("hebo")
}\if{html}{\out{</div>}}
}

\section{Parameters}{

\describe{
\item{\code{n_suggestions}}{\code{integer(1)}\cr
Number of candidate configurations to suggest per iteration.
Default is \code{1L}.}
\item{\code{n_init}}{\code{integer(1)}\cr
Number of random initialization points before the surrogate model is fitted.
Default is \verb{1 + number of parameters}.}
\item{\code{seed}}{\code{integer(1)}\cr
Random seed for the scramble sequence.
Default is \code{NULL} (no fixed seed).}
\item{\code{surrogate}}{\code{character(1)}\cr
Surrogate model to use.
\code{"gp"} (Gaussian Process) or \code{"rf"} (Random Forest).
Default is \code{"gp"}.}
\item{\code{rf_n_estimators}}{\code{integer(1)}\cr
Number of trees in the Random Forest surrogate.
Only used when \code{surrogate = "rf"}.
Default is \code{20L}.}
\item{\code{gp_lr}}{\code{numeric(1)}\cr
Learning rate for the GP model.
Only used when \code{surrogate = "gp"}.
Default is \code{0.01}.}
\item{\code{gp_num_epochs}}{\code{integer(1)}\cr
Number of training epochs for the GP model.
Only used when \code{surrogate = "gp"}.
Default is \code{100L}.}
\item{\code{gp_noise_free}}{\code{logical(1)}\cr
Whether to treat the objective as noise-free.
Only used when \code{surrogate = "gp"}.
Default is \code{FALSE}.}
\item{\code{gp_noise_lb}}{\code{numeric(1)}\cr
Lower bound on GP noise variance.
Only used when \code{surrogate = "gp"}.
Default is \code{8e-4}.}
\item{\code{gp_pred_likeli}}{\code{logical(1)}\cr
Whether to use predictive likelihood.
Only used when \code{surrogate = "gp"}.
Default is \code{FALSE}.}
\item{\code{acq_function}}{\code{character(1)}\cr
Acquisition function to use.
\code{"mace"} (Multi-Objective Acquisition Criterion Ensemble) or \code{"lcb"} (Lower Confidence Bound).
Default is \code{"mace"}.}
\item{\code{es}}{\code{character(1)}\cr
Evolutionary strategy for acquisition function optimization.
One of \code{"ga"}, \code{"brkga"}, \code{"de"}, \code{"nelder-mead"}, \code{"pattern-search"}, \code{"cmaes"}, \code{"pso"},
\code{"nsga2"}, \code{"rnsga2"}, \code{"nsga3"}, \code{"unsga3"}, \code{"rnsga3"}, \code{"moead"}, \code{"ctaea"}.
Default is \code{"nsga2"}.}
}
}

\section{Progress Bars}{

\verb{$optimize()} supports progress bars via the package \CRANpkg{progressr}
combined with a \link{Terminator}. Simply wrap the function in
\code{progressr::with_progress()} to enable them. We recommend to use package
\CRANpkg{progress} as backend; enable with \code{progressr::handlers("progress")}.
}

\section{Super classes}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{\link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch}} -> \code{OptimizerBatchHEBO}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerBatchHEBO-new}{\code{OptimizerBatchHEBO$new()}}
\item \href{#method-OptimizerBatchHEBO-clone}{\code{OptimizerBatchHEBO$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimizerBatch" data-id="optimize"><a href='../../bbotk/html/OptimizerBatch.html#method-OptimizerBatch-optimize'><code>bbotk::OptimizerBatch$optimize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchHEBO-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchHEBO-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchHEBO$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchHEBO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchHEBO-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchHEBO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
