% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatchOptuna.R
\name{mlr_optimizers_optuna}
\alias{mlr_optimizers_optuna}
\alias{OptimizerBatchOptuna}
\title{Optimization via Optuna}
\description{
\code{OptimizerBatchOptuna} class that implements optimization via the Python
Optuna framework. Calls Optuna via \CRANpkg{reticulate}.
}
\section{Dictionary}{

This \link{Optimizer} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_optimizers} or with the associated sugar function \code{\link[=opt]{opt()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_optimizers$get("optuna")
opt("optuna")
}\if{html}{\out{</div>}}
}

\section{Parameters}{

\describe{
\item{\code{sampler}}{\code{character(1)}\cr
Optuna sampler to use.
One of \code{"tpe"} (Tree-structured Parzen Estimator), \code{"gp"} (Gaussian Process),
\code{"nsga2"} (NSGA-II), \code{"nsga3"} (NSGA-III), \code{"bruteforce"}, or \code{"random"}.
\code{"nsga2"} and \code{"nsga3"} are recommended for multi-criteria optimization.
Default is \code{"tpe"}.}
\item{\code{n_startup_trials}}{\code{integer(1)}\cr
Number of random startup trials before the TPE sampler model is fitted.
Only used when \code{sampler = "tpe"}.
Default is \code{10L}.}
\item{\code{seed}}{\code{integer(1)}\cr
Random seed passed to the sampler for reproducibility.
Default is \code{NULL} (no fixed seed).}
}
}

\section{Progress Bars}{

\verb{$optimize()} supports progress bars via the package \CRANpkg{progressr}
combined with a \link{Terminator}. Simply wrap the function in
\code{progressr::with_progress()} to enable them. We recommend to use package
\CRANpkg{progress} as backend; enable with \code{progressr::handlers("progress")}.
}

\section{Super classes}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{\link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch}} -> \code{OptimizerBatchOptuna}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerBatchOptuna-new}{\code{OptimizerBatchOptuna$new()}}
\item \href{#method-OptimizerBatchOptuna-clone}{\code{OptimizerBatchOptuna$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimizerBatch" data-id="optimize"><a href='../../bbotk/html/OptimizerBatch.html#method-OptimizerBatch-optimize'><code>bbotk::OptimizerBatch$optimize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchOptuna-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchOptuna-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchOptuna$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchOptuna-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchOptuna-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchOptuna$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
