% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorClockTime.R
\name{mlr_terminators_clock_time}
\alias{mlr_terminators_clock_time}
\alias{TerminatorClockTime}
\title{Terminator that stops according to the clock time}
\arguments{
\item{archive}{(\link{Archive}).}
}
\value{
\code{logical(1)}.
}
\description{
Class to terminate the optimization after a fixed time point has been reached (as
reported by \code{\link[=Sys.time]{Sys.time()}}).

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("clock_time")
term("clock_time")
}
}

\section{Parameters}{

\itemize{
\item \code{stop_time} (\code{POSIXct(1)})\cr
Terminator stops after this point in time.
Mutually exclusive with argument \code{secs}.
}
}

\examples{
stop_time = as.POSIXct("2030-01-01 00:00:00")
term("clock_time", stop_time = stop_time)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_run_time}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
