% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvals.R
\name{mlr_terminators_evals}
\alias{mlr_terminators_evals}
\alias{TerminatorEvals}
\title{Terminator that stops after a number of evaluations}
\arguments{
\item{archive}{(\link{Archive}).}
}
\value{
\code{logical(1)}.
}
\description{
Class to terminate the optimization depending on the number of evaluations.
An evaluation is defined by one resampling of a parameter value.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("evals")
term("evals")
}
}

\section{Parameters}{

\itemize{
\item \code{n_evals} (\code{integer(1)})\cr
Number of allowed evaluations, default is 100L
}
}

\examples{
TerminatorEvals$new()
term("evals", n_evals = 5)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_run_time}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
