% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorModelTime.R
\name{mlr_terminators_model_time}
\alias{mlr_terminators_model_time}
\alias{TerminatorModelTime}
\title{Terminator that stops after a budget of model evaluation time is depleted}
\arguments{
\item{instance}{(\link{TuningInstance}).}
}
\value{
\code{logical(1)}.
}
\description{
Class to terminate the optimization after a given model evaluation budget is
exceeded. The terminator measures the used time to train and predict all
models contained in the archive.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("model_time")
term("model_time")
}
}

\section{Parameters}{

\itemize{
\item \code{secs} (\code{numeric(1)})\cr
Maximum allowed time, in seconds, default is 0.
}
}

\examples{
TerminatorModelTime$new()
term("model_time", secs = 10 * 3600)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
