% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRunTime.R
\name{mlr_terminators_run_time}
\alias{mlr_terminators_run_time}
\alias{TerminatorRunTime}
\title{Terminator that stops according to the run time}
\arguments{
\item{archive}{(\link{Archive}).}
}
\value{
\code{logical(1)}.
}
\description{
Class to terminate the optimization after the complete process took a
number of seconds on the clock.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("run_time")
term("run_time")
}
}

\section{Parameters}{

\itemize{
\item \code{secs} (\code{numeric(1)})\cr
Maximum allowed time, in seconds, default is 100.
Mutually exclusive with argument \code{stop_time}.
}
}

\examples{
term("run_time", secs = 1800)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
