% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorStagnation.R
\name{mlr_terminators_stagnation}
\alias{mlr_terminators_stagnation}
\alias{TerminatorStagnation}
\title{Terminator that stops when optimization does not improve}
\arguments{
\item{instance}{(\link{TuningInstance}).}
}
\value{
\code{logical(1)}.
}
\description{
Class to terminate the optimization after the performance stagnates, i.e.
does not improve more than \code{threshold} over the last \code{iters} iterations.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("stagnation")
term("stagnation")
}
}

\section{Parameters}{

\itemize{
\item \code{iters} (\code{integer(1)})\cr
Number of iterations to evaluate the performance improvement on, default is 10.
\item \code{threshold} (\code{numeric(1)})\cr
If the improvement is less than \code{threshold}, optimization is stopped, default is \code{0}.
}
}

\examples{
TerminatorStagnation$new()
term("stagnation", iters = 5, threshold = 1e-5)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_model_time}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
