% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_test_functions.R
\docType{data}
\name{mlr_test_functions}
\alias{mlr_test_functions}
\title{Dictionary of Optimization Test Functions}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing well-known optimization test functions
as \link{ObjectiveRFun} objects.

Each objective has two additional fields beyond the standard \link{ObjectiveRFun} interface:
\itemize{
\item \code{optimum} (\code{numeric(1)}) - Known global optimum value (f*).
\item \code{optimum_x} (\code{list()}) - List of known global optima (each a named list).
}

For a more convenient way to retrieve test functions, see \code{\link[=otfun]{otfun()}}/\code{\link[=otfuns]{otfuns()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict, ..., objects = FALSE)}\cr
\link[mlr3misc:Dictionary]{mlr3misc::Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{\link[data.table:data.table]{data.table::data.table()}} with fields "key", "label", "optimum",
and "optimum_x" as columns.
If \code{objects} is set to \code{TRUE}, the constructed objects are returned in the list column named \code{object}.
}
}

\examples{
as.data.table(mlr_test_functions)
obj = mlr_test_functions$get("branin")
obj$eval(list(x1 = 0, x2 = 0))
}
\seealso{
Sugar functions: \code{\link[=otfun]{otfun()}}, \code{\link[=otfuns]{otfuns()}}
}
\keyword{datasets}
