% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mult_max_to_min}
\alias{mult_max_to_min}
\title{Multiplication vector for output}
\usage{
mult_max_to_min(codomain)
}
\arguments{
\item{codomain}{\link{ParamSet}}
}
\value{
'numeric()'
}
\description{
Returns a numeric vector with values -1 and 1.
If you multiply this vector with an outcome of \code{codomain} it will be turned into a minimization problem.
}
\keyword{internal}
