% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_selection.R
\name{nds_selection}
\alias{nds_selection}
\title{Best points w.r.t. non dominated sorting with hypervolume contribution.}
\usage{
nds_selection(points, n_select, ref_point = NULL, minimize = TRUE)
}
\arguments{
\item{points}{(\code{matrix()})\cr
Numeric matrix with each column corresponding to a point}

\item{n_select}{(\code{integer(1L)})\cr
Amount of points to select.}

\item{ref_point}{(\code{numeric()})\cr
Reference point for hypervolume.}

\item{minimize}{('logical()')\cr
Should the ranking be based on minimization?
Must be of length 1 for all dimensions or of length \code{nrow(points)} for each dimension.
Default is \code{TRUE} for each dimension.}
}
\value{
Vector of indices of selected points
}
\description{
Select best subset of points by non dominated sorting with
hypervolume contribution for tie breaking. Works on an arbitrary dimension
of size two or higher.
Non-dominated sorting is computed with \code{\link[moocore:pareto_rank]{moocore::pareto_rank()}} and
hypervolume contributions with \code{\link[moocore:hv_contributions]{moocore::hv_contributions()}}.
Boundary points, i.e., points that are best in at least one objective within their front,
always survive tie breaking.
}
\keyword{internal}
