% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Niches.R
\name{Niches}
\alias{Niches}
\title{Niches}
\description{
Describes a set of niches related to a feature function in QDO.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{(\code{character(1)}).}

\item{\code{niches}}{(\code{list()}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Niches$new()}}
\item \href{#method-get_niche}{\code{Niches$get_niche()}}
\item \href{#method-get_niche_many}{\code{Niches$get_niche_many()}}
\item \href{#method-get_niche_dt}{\code{Niches$get_niche_dt()}}
\item \href{#method-clone}{\code{Niches$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Niches$new(id, niches)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}).}

\item{\code{niches}}{(\code{list()}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_niche"></a>}}
\if{latex}{\out{\hypertarget{method-get_niche}{}}}
\subsection{Method \code{get_niche()}}{
Get the niche a point belongs to based on the evaluation of the feature function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Niches$get_niche(gval)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gval}}{(\code{list()})\cr
A \code{list()} that contains a value of the feature function, e.g., \code{list(g1 = 1, g2 = 2)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)} indicating the niche.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_niche_many"></a>}}
\if{latex}{\out{\hypertarget{method-get_niche_many}{}}}
\subsection{Method \code{get_niche_many()}}{
Get the niches multiple points belong to based on the evaluations of the feature function stored in a \code{list()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Niches$get_niche_many(gvals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gvals}}{(\code{list()})\cr
A \code{list()} that contains values of the feature function for multiple points, e.g., \code{list(list(g1 = 1, g2 = 2), list(g1 = 2, g2 = 3))}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character()} indicating the niches.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_niche_dt"></a>}}
\if{latex}{\out{\hypertarget{method-get_niche_dt}{}}}
\subsection{Method \code{get_niche_dt()}}{
Get the niches multiple points belong to based on the evaluations of the feature function stored in a \link[data.table:data.table]{data.table::data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Niches$get_niche_dt(gvaldt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gvaldt}}{(\link[data.table:data.table]{data.table::data.table})\cr
A \link[data.table:data.table]{data.table::data.table} containing values of the feature function for multiple points as rows.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\link[data.table:data.table]{data.table::data.table} containing the niches.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Niches$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
