% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Niches.R
\name{NichesBoundaries}
\alias{NichesBoundaries}
\title{NichesBoundaries}
\description{
Describes a set of niches related to a feature function in QDO based on some boundaries.
}
\section{Super class}{
\code{\link[bbotk:Niches]{bbotk::Niches}} -> \code{NichesBoundaries}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NichesBoundaries$new()}}
\item \href{#method-clone}{\code{NichesBoundaries$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Niches" data-id="get_niche">}\href{../../bbotk/html/Niches.html#method-get_niche}{\code{bbotk::Niches$get_niche()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Niches" data-id="get_niche_dt">}\href{../../bbotk/html/Niches.html#method-get_niche_dt}{\code{bbotk::Niches$get_niche_dt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Niches" data-id="get_niche_many">}\href{../../bbotk/html/Niches.html#method-get_niche_many}{\code{bbotk::Niches$get_niche_many()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NichesBoundaries$new(id, niches_boundaries)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}).}

\item{\code{niches_boundaries}}{(\code{list()}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NichesBoundaries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
