% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveRFun.R
\name{ObjectiveRFun}
\alias{ObjectiveRFun}
\title{Objective interface where user cass pass R function}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{

}
\section{Construction}{
\preformatted{obj = ObjectiveRFun$new(fun, domain, codomain = NULL, id = "function", properties = character(0))
}
\itemize{
\item \code{objective} :: \link{Objective}.
\item \code{param_set} :: \link{ParamSet}.
\item \code{terminator} :: \link{Terminator}.
}
}

