% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{opt}
\alias{opt}
\title{Syntactic Sugar Optimizer Construction}
\usage{
opt(.key, ...)

opt(.key, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}
}
\value{
\itemize{
\item \link{Optimizer} for \code{opt()}.
\item list of \link{Optimizer} for \code{opts()}.
}
}
\description{
This function complements \link{mlr_optimizer} with functions in the spirit
of \link[mlr3:mlr_sugar]{mlr3::mlr_sugar}.
}
\examples{
opt("random_search", batch_size = 10)
}
