% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimInstanceRush.R
\name{OptimInstanceRush}
\alias{OptimInstanceRush}
\title{Optimization Instance with Rush}
\description{
Abstract base class for \link{OptimInstanceRushSingleCrit} and \link{OptimInstanceRushMultiCrit}.
The optimization instances specify an optimization problem for \link{Optimizer}s.
Points are evaluated asynchronously with the \code{rush} package.
The function \code{\link[=oi]{oi()}} creates an \link{OptimInstanceRushSingleCrit} or \link{OptimInstanceRushMultiCrit} and the function \code{\link[=bb_optimize]{bb_optimize()}} creates an instance internally.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objective}}{(\link{Objective})\cr
Objective function of the instance.}

\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specification of the search space for the \link{Optimizer}.}

\item{\code{terminator}}{\link{Terminator}\cr
Termination criterion of the optimization.}

\item{\code{rush}}{(\code{Rush})\cr
Rush controller for parallel optimization.}

\item{\code{callbacks}}{(list of \link[mlr3misc:Callback]{mlr3misc::Callback})\cr
Callbacks applied during the optimization.}

\item{\code{archive}}{(\link{ArchiveRush})\cr
Contains all performed function calls of the Objective.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{(\link[data.table:data.table]{data.table::data.table})\cr
Get result}

\item{\code{result_x_search_space}}{(\link[data.table:data.table]{data.table::data.table})\cr
x part of the result in the \emph{search space}.}

\item{\code{result_x_domain}}{(\code{list()})\cr
(transformed) x part of the result in the \emph{domain space} of the objective.}

\item{\code{result_y}}{(\code{numeric()})\cr
Optimal outcome.}

\item{\code{is_terminated}}{(\code{logical(1)}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimInstanceRush-new}{\code{OptimInstanceRush$new()}}
\item \href{#method-OptimInstanceRush-format}{\code{OptimInstanceRush$format()}}
\item \href{#method-OptimInstanceRush-print}{\code{OptimInstanceRush$print()}}
\item \href{#method-OptimInstanceRush-start_workers}{\code{OptimInstanceRush$start_workers()}}
\item \href{#method-OptimInstanceRush-eval_async}{\code{OptimInstanceRush$eval_async()}}
\item \href{#method-OptimInstanceRush-clear}{\code{OptimInstanceRush$clear()}}
\item \href{#method-OptimInstanceRush-clone}{\code{OptimInstanceRush$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$new(
  objective,
  search_space = NULL,
  terminator,
  rush,
  callbacks = list(),
  archive = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(\link{Objective})\cr
Objective function.}

\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies the search space for the \link{Optimizer}. The \link[paradox:ParamSet]{paradox::ParamSet}
describes either a subset of the \code{domain} of the \link{Objective} or it describes
a set of parameters together with a \code{trafo} function that transforms values
from the search space to values of the domain. Depending on the context, this
value defaults to the domain of the objective.}

\item{\code{terminator}}{\link{Terminator}\cr
Termination criterion.}

\item{\code{rush}}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}

\item{\code{callbacks}}{(list of \link[mlr3misc:Callback]{mlr3misc::Callback})\cr
List of callbacks.}

\item{\code{archive}}{(\link{Archive}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-format"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-format}{}}}
\subsection{Method \code{format()}}{
Helper for print outputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$format(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-print"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-start_workers"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-start_workers}{}}}
\subsection{Method \code{start_workers()}}{
Start workers with the \code{future} package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$start_workers(
  n_workers = NULL,
  packages = NULL,
  host = "local",
  heartbeat_period = NULL,
  heartbeat_expire = NULL,
  lgr_thresholds = NULL,
  await_workers = TRUE,
  detect_lost_tasks = FALSE,
  freeze_archive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_workers}}{(\code{integer(1)})\cr
Number of workers to be started.
If \code{NULL} the maximum number of free workers is used.}

\item{\code{packages}}{(\code{character()})\cr
Packages to be loaded by the workers.}

\item{\code{host}}{(\code{character(1)})\cr
Worker is started on a local or remote host.}

\item{\code{heartbeat_period}}{(\code{integer(1)})\cr
Period of the heartbeat in seconds.}

\item{\code{heartbeat_expire}}{(\code{integer(1)})\cr
Time to live of the heartbeat in seconds.}

\item{\code{lgr_thresholds}}{(named \code{character()} or \code{numeric()})\cr
Thresholds for the \code{lgr} package e.g. c(rush = "debug", bbotk = "info").
If \code{NULL}, no log messages from the workers are saved.}

\item{\code{await_workers}}{(\code{logical(1)})\cr
Whether to wait until all workers are available.}

\item{\code{detect_lost_tasks}}{(\code{logical(1)})\cr
Whether to detect lost tasks.
Defaults to \code{FALSE}.}

\item{\code{freeze_archive}}{(\code{logical(1)})\cr
If \code{TRUE}, the archive is copied from redis to a local data.table after tuning.
This is helpful when the tuning is run on a remote machine and the archive is serialized.
Only used if \code{rush} is supplied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-eval_async"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-eval_async}{}}}
\subsection{Method \code{eval_async()}}{
Adds points in \code{xdt} to the queue.
The points are evaluated by calling the \link{Objective} asynchronously.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$eval_async(xdt, wait = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{(\code{data.table::data.table()})\cr
x values as \code{data.table()} with one point per row.
Contains the value in  the \emph{search space} of the \link{OptimInstance} object.
Can contain additional columns for extra information.}

\item{\code{wait}}{(\code{logical(1)})\cr
If \code{TRUE}, wait for all evaluations to finish.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-clear"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-clear}{}}}
\subsection{Method \code{clear()}}{
Reset terminator and clear all evaluation results from archive and results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRush-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRush-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRush$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
