% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimInstanceRushSingleCrit.R
\name{OptimInstanceRushSingleCrit}
\alias{OptimInstanceRushSingleCrit}
\title{Single Criterion Optimization Instance with Rush}
\description{
The \link{OptimInstanceRushSingleCrit} specifies an optimization problem for \link{Optimizer}s.
Points are evaluated asynchronously with the \code{rush} package.
The function \code{\link[=oi]{oi()}} creates an \link{OptimInstanceRushSingleCrit} and the function \code{\link[=bb_optimize]{bb_optimize()}} creates an instance internally.
}
\section{Super class}{
\code{\link[bbotk:OptimInstanceRush]{bbotk::OptimInstanceRush}} -> \code{OptimInstanceRushSingleCrit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimInstanceRushSingleCrit-new}{\code{OptimInstanceRushSingleCrit$new()}}
\item \href{#method-OptimInstanceRushSingleCrit-clone}{\code{OptimInstanceRushSingleCrit$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceRush" data-id="clear"><a href='../../bbotk/html/OptimInstanceRush.html#method-OptimInstanceRush-clear'><code>bbotk::OptimInstanceRush$clear()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceRush" data-id="eval_async"><a href='../../bbotk/html/OptimInstanceRush.html#method-OptimInstanceRush-eval_async'><code>bbotk::OptimInstanceRush$eval_async()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceRush" data-id="format"><a href='../../bbotk/html/OptimInstanceRush.html#method-OptimInstanceRush-format'><code>bbotk::OptimInstanceRush$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceRush" data-id="print"><a href='../../bbotk/html/OptimInstanceRush.html#method-OptimInstanceRush-print'><code>bbotk::OptimInstanceRush$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstanceRush" data-id="start_workers"><a href='../../bbotk/html/OptimInstanceRush.html#method-OptimInstanceRush-start_workers'><code>bbotk::OptimInstanceRush$start_workers()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRushSingleCrit-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRushSingleCrit-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRushSingleCrit$new(
  objective,
  search_space = NULL,
  terminator,
  rush,
  callbacks = list(),
  archive = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{objective}}{(\link{Objective})\cr
Objective function.}

\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies the search space for the \link{Optimizer}. The \link[paradox:ParamSet]{paradox::ParamSet}
describes either a subset of the \code{domain} of the \link{Objective} or it describes
a set of parameters together with a \code{trafo} function that transforms values
from the search space to values of the domain. Depending on the context, this
value defaults to the domain of the objective.}

\item{\code{terminator}}{\link{Terminator}\cr
Termination criterion.}

\item{\code{rush}}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}

\item{\code{callbacks}}{(list of \link[mlr3misc:Callback]{mlr3misc::Callback})\cr
List of callbacks.}

\item{\code{archive}}{(\link{Archive}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimInstanceRushSingleCrit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimInstanceRushSingleCrit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceRushSingleCrit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
