% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{optimize_async_default}
\alias{optimize_async_default}
\title{Default Asynchronous Optimization}
\usage{
optimize_async_default(
  instance,
  optimizer,
  design = NULL,
  n_workers = NULL,
  profiles = NULL
)
}
\arguments{
\item{instance}{\link{OptimInstanceAsync}.}

\item{optimizer}{\link{OptimizerAsync}.}

\item{design}{\code{\link[data.table:data.table]{data.table::data.table()}}\cr
(Initial) design send to the queue.}

\item{n_workers}{Number of workers to be started.
Defaults to the number of workers set by \code{\link[rush:rush_plan]{rush::rush_plan()}}.}

\item{profiles}{(named \code{integer()})\cr
Number of workers to be started on each \CRANpkg{mirai} compute profile, e.g. \code{c(cpu = 2, gpu = 2)}.
Defaults to the profiles set by \code{\link[rush:rush_plan]{rush::rush_plan()}}.
Cannot be combined with \code{n_workers}.}
}
\description{
Used internally in \link{OptimizerAsync}.
}
\keyword{internal}
