% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatch.R
\name{optimize_batch_default}
\alias{optimize_batch_default}
\title{Default Batch Optimization Function}
\usage{
optimize_batch_default(instance, optimizer)
}
\arguments{
\item{instance}{\link{OptimInstance}}

\item{optimizer}{\link{OptimizerBatch}}
}
\value{
\link[data.table:data.table]{data.table::data.table}
}
\description{
Used internally in the \link{OptimizerBatch}.
}
\keyword{internal}
