% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{optimize_decentralized}
\alias{optimize_decentralized}
\title{Decentralized Optimization Function}
\usage{
optimize_decentralized(inst, self, privat)
}
\arguments{
\item{inst}{\link{OptimInstance}}

\item{self}{\link{Optimizer}}

\item{private}{(\code{environment()})}
}
\value{
\link[data.table:data.table]{data.table::data.table}
}
\description{
Used internally in the \link{Optimizer}.
Brings together the private \code{.optimize()} method and the private \code{.assign_result()} method.
}
\keyword{internal}
