% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{optimize_default}
\alias{optimize_default}
\title{Default optimization function}
\usage{
optimize_default(inst, self, private)
}
\arguments{
\item{inst}{\link{OptimInstance}}

\item{self}{\link{Optimizer}}

\item{private}{(\code{environment()})}
}
\description{
Default optimization function
}
\keyword{internal}
