% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{otfun}
\alias{otfun}
\alias{otfuns}
\title{Syntactic Sugar for Optimization Test Functions}
\usage{
otfun(.key, ...)

otfuns(.keys, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve
the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor. See
\code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}

\item{.keys}{(\code{character()})\cr
Keys passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve
multiple objects.}
}
\value{
\itemize{
\item \link{ObjectiveRFun} for \code{otfun()}.
\item list of \link{ObjectiveRFun} for \code{otfuns()}.
}
}
\description{
This function complements \link{mlr_test_functions} with functions in the spirit
of \code{mlr_sugar} from \CRANpkg{mlr3}.
}
\examples{
obj = otfun("branin")
obj$eval(list(x1 = 1, x2 = 2))
}
