% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{search_start}
\alias{search_start}
\title{Get start values for optimizers.}
\usage{
search_start(search_space, type = "random")
}
\arguments{
\item{search_space}{\link{ParamSet}.}

\item{type}{(\code{character(1)})\cr
\code{random} start values or \code{center} of search space?}
}
\value{
named 'numeric()'
}
\description{
Returns a named numeric vector with start
values for optimizers.
}
\keyword{internal}
