% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatchFocusSearch.R
\name{shrink_ps}
\alias{shrink_ps}
\title{Shrink a ParamSet towards a point.}
\usage{
shrink_ps(param_set, x, check.feasible = FALSE)
}
\arguments{
\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
The \link[paradox:ParamSet]{paradox::ParamSet} to be shrunk.}

\item{x}{(\link[data.table:data.table]{data.table::data.table})\cr
\link[data.table:data.table]{data.table::data.table} with one row containing the point to shrink
around.}

\item{check.feasible}{(\code{logical(1)})\cr
Should feasibility of the parameters be checked?
If feasibility is not checked, and invalid values are present, no shrinking
will be done.
Default is \code{FALSE}.}
}
\value{
\link[paradox:ParamSet]{paradox::ParamSet}
}
\description{
Shrinks a \link[paradox:ParamSet]{paradox::ParamSet} towards a point.
Boundaries of numeric values are shrunk to an interval around the point of
half of the previous length, while for discrete variables, a random
(currently not chosen) level is dropped.

Note that for \code{\link[paradox:p_lgl]{paradox::p_lgl()}}s the value to be shrunk around is set as
the \code{default} value instead of dropping a level. Also, a tag \code{shrunk} is
added.

Note that the returned \link[paradox:ParamSet]{paradox::ParamSet} has lost all its original
\code{default}s, as they may have become infeasible.

\code{x} is always expected to contain the values on the scale of \code{param_set},
i.e. the untransformed values before a trafo is applied.
}
\examples{
library(paradox)
library(data.table)
param_set = ps(
  x1 = p_dbl(lower = 0, upper = 10),
  x2 = p_int(lower = -10, upper = 10),
  x3 = p_fct(levels = c("a", "b", "c")),
  x4 = p_lgl()
)
x = data.table(x1 = 5, x2 = 0, x3 = "b", x4 = FALSE)
shrink_ps(param_set, x = x)
}
