% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{start_async_optimize}
\alias{start_async_optimize}
\title{Start Decentralized Optimization}
\usage{
start_async_optimize(inst, self, private, context)
}
\arguments{
\item{inst}{\link{OptimInstance}}

\item{self}{\link{Optimizer}}

\item{private}{(\code{environment()})}
}
\description{
Used internally in \link{OptimizerAsync}.
Starts the workers and pushes the necessary objects to the workers.
}
\keyword{internal}
