% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{terminated_error}
\alias{terminated_error}
\title{Termination Error}
\usage{
terminated_error(optim_instance)
}
\arguments{
\item{optim_instance}{\link{OptimInstance}\cr
OptimInstance that terminated.}
}
\value{
A \code{'terminated_error'} condition (inherits from \code{'error'} and \code{'condition'}).
}
\description{
Error class for termination.
}
