% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{tiny_logging}
\alias{tiny_logging}
\title{Tiny Logging}
\usage{
tiny_logging(instance, optimizer)
}
\arguments{
\item{instance}{\link{OptimInstanceAsync}.}

\item{optimizer}{\link{OptimizerAsync}.}
}
\description{
Used internally in \link{OptimizerAsync}.
Adapts tiny logging to the different instance types.
}
\examples{
options(bbotk.tiny_logging = TRUE)
}
\keyword{internal}
