% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{tiny_result}
\alias{tiny_result}
\title{Tiny Result}
\usage{
tiny_result(instance, optimizer)
}
\arguments{
\item{instance}{\link{OptimInstanceAsync}.}

\item{optimizer}{\link{OptimizerAsync}.}
}
\description{
Used internally in \link{OptimizerAsync}.
Adapts tiny result to the different instance types.
}
\keyword{internal}
