% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{trafo_xs}
\alias{trafo_xs}
\title{Calculate the transformed x-values}
\usage{
trafo_xs(xs, search_space)
}
\arguments{
\item{xs}{(\code{list()}) \cr
List of x-values.}

\item{search_space}{\link[paradox:ParamSet]{paradox::ParamSet}\cr
Search space.}
}
\description{
Transforms a given \code{list()} to a `list()`` with transformed x values.
}
