% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{transform_xdt_to_xss}
\alias{transform_xdt_to_xss}
\title{Calculates the transformed x-values}
\usage{
transform_xdt_to_xss(xdt, search_space)
}
\arguments{
\item{xdt}{(\code{data.table}) \cr
The data table with x-colums.
Column names have to match ids of the \code{search_space}.
However, \code{xdt} can contain additinal columns.}

\item{search_space}{\link[paradox:ParamSet]{paradox::ParamSet} \cr
The ParamSet.}
}
\description{
Transforms a given \code{data.table} to a list with transformed x values.
If no trafo is defined it will just convert the \code{data.table} to a list.
Mainly for internal usage.
}
\keyword{internal}
