% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{trm}
\alias{trm}
\alias{trms}
\title{Syntactic Sugar Terminator Construction}
\usage{
trm(.key, ...)

trms(.keys, ...)
}
\arguments{
\item{.key}{(\code{character(1)})\cr
Key passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve
the object.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to be set as parameters in the
\link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field. See
\code{\link[mlr3misc:dictionary_sugar_get]{mlr3misc::dictionary_sugar_get()}} for more details.}

\item{.keys}{(\code{character()})\cr
Keys passed to the respective \link[mlr3misc:Dictionary]{dictionary} to retrieve
multiple objects.}
}
\value{
\itemize{
\item \link{Terminator} for \code{trm()}.
\item list of \link{Terminator} for \code{trms()}.
}
}
\description{
This function complements \link{mlr_terminators} with functions in the spirit
of \code{mlr_sugar} from \CRANpkg{mlr3}.
}
\examples{
trm("evals", n_evals = 10)
}
