% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{wait_for_async_optimize}
\alias{wait_for_async_optimize}
\title{Wait for Decentralized Optimization}
\usage{
wait_for_async_optimize(inst, self, private)
}
\arguments{
\item{inst}{\link{OptimInstance}}

\item{self}{\link{Optimizer}}

\item{private}{(\code{environment()})}
}
\description{
Used internally in \link{OptimizerAsync}.
Waits until optimization is finished.
Prints log messages from the workers and checks for errors.
}
\keyword{internal}
