% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Archive.R
\name{Archive}
\alias{Archive}
\title{Container for objective function evaluations}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Container around a data.table which stores all performed \link{Objective} function calls by the \link{Evaluator}.
}
\section{Construction}{
\preformatted{archive = Archive$new(domain, codomain)
}
\itemize{
\item \code{domain} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
Domain of objective function that is logged into archive.
\item \code{codomain} :: \link[paradox:ParamSet]{paradox::ParamSet}\cr
Codomain of objective function that is logged into archive.
}
}

\section{Fields}{

\itemize{
\item \code{domain} :: \link[paradox:ParamSet]{paradox::ParamSet} from construction\cr
\item \code{codomain} :: \link[paradox:ParamSet]{paradox::ParamSet} from construction\cr
\item \code{data} :: \link[data.table:data.table]{data.table::data.table}\cr
Holds data of the archive.
\item \code{n_evals} :: \code{integer(1)}\cr
Number of evaluations stored in the container.
}
}

\section{Methods}{

\itemize{
\item \code{add_evals(dt)}\cr
\link[data.table:data.table]{data.table::data.table} -> \code{self}\cr
Adds function evaluations to the archive table.
}
}

