% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAsync.R
\name{ArchiveAsync}
\alias{ArchiveAsync}
\title{Rush Data Storage}
\description{
The \code{ArchiveAsync} stores all evaluated points and performance scores in a \link[rush:Rush]{rush::Rush} data base.
}
\section{S3 Methods}{

\itemize{
\item \code{as.data.table(archive)}\cr
\link{ArchiveAsync} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a tabular view of all performed function calls of the Objective.
The \code{x_domain} column is unnested to separate columns.
}
}

\examples{
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "mirai")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("async_random_search")

# trigger optimization
optimizer$optimize(instance)

# all evaluated configuration
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)

# reset the rush data base
instance$rush$reset()
}
}
\section{Super class}{
\code{\link[bbotk:Archive]{Archive}} -> \code{ArchiveAsync}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{rush}}{(\code{Rush})\cr
Rush controller for parallel optimization.}
  }
  \if{html}{\out{</div>}}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all finished points.}

    \item{\code{queued_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all queued points.}

    \item{\code{running_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all running points.}

    \item{\code{finished_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all finished points.}

    \item{\code{failed_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all failed points.}

    \item{\code{n_queued}}{(\code{integer(1)})\cr
Number of queued points.}

    \item{\code{n_running}}{(\code{integer(1)})\cr
Number of running points.}

    \item{\code{n_finished}}{(\code{integer(1)})\cr
Number of finished points.}

    \item{\code{n_failed}}{(\code{integer(1)})\cr
Number of failed points.}

    \item{\code{n_evals}}{(\code{integer(1)})\cr
Number of evaluations stored in the archive.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ArchiveAsync-initialize}{\code{ArchiveAsync$new()}}
    \item \href{#method-ArchiveAsync-push_points}{\code{ArchiveAsync$push_points()}}
    \item \href{#method-ArchiveAsync-push_point}{\code{ArchiveAsync$push_point()}}
    \item \href{#method-ArchiveAsync-push_running_points}{\code{ArchiveAsync$push_running_points()}}
    \item \href{#method-ArchiveAsync-push_running_point}{\code{ArchiveAsync$push_running_point()}}
    \item \href{#method-ArchiveAsync-push_finished_points}{\code{ArchiveAsync$push_finished_points()}}
    \item \href{#method-ArchiveAsync-push_finished_point}{\code{ArchiveAsync$push_finished_point()}}
    \item \href{#method-ArchiveAsync-push_failed_points}{\code{ArchiveAsync$push_failed_points()}}
    \item \href{#method-ArchiveAsync-push_failed_point}{\code{ArchiveAsync$push_failed_point()}}
    \item \href{#method-ArchiveAsync-pop_point}{\code{ArchiveAsync$pop_point()}}
    \item \href{#method-ArchiveAsync-finish_points}{\code{ArchiveAsync$finish_points()}}
    \item \href{#method-ArchiveAsync-finish_point}{\code{ArchiveAsync$finish_point()}}
    \item \href{#method-ArchiveAsync-fail_points}{\code{ArchiveAsync$fail_points()}}
    \item \href{#method-ArchiveAsync-fail_point}{\code{ArchiveAsync$fail_point()}}
    \item \href{#method-ArchiveAsync-push_result}{\code{ArchiveAsync$push_result()}}
    \item \href{#method-ArchiveAsync-data_with_state}{\code{ArchiveAsync$data_with_state()}}
    \item \href{#method-ArchiveAsync-best}{\code{ArchiveAsync$best()}}
    \item \href{#method-ArchiveAsync-nds_selection}{\code{ArchiveAsync$nds_selection()}}
    \item \href{#method-ArchiveAsync-clear}{\code{ArchiveAsync$clear()}}
    \item \href{#method-ArchiveAsync-clone}{\code{ArchiveAsync$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="format"><a href='../../bbotk/html/Archive.html#method-Archive-format'><code>Archive$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="help"><a href='../../bbotk/html/Archive.html#method-Archive-help'><code>Archive$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="print"><a href='../../bbotk/html/Archive.html#method-Archive-print'><code>Archive$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-initialize}{}}}
\subsection{\code{ArchiveAsync$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$new(search_space, codomain, check_values = FALSE, rush)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies the search space for the \link{Optimizer}. The \link[paradox:ParamSet]{paradox::ParamSet}
describes either a subset of the \code{domain} of the \link{Objective} or it describes
a set of parameters together with a \code{trafo} function that transforms values
from the search space to values of the domain. Depending on the context, this
value defaults to the domain of the objective.}
      \item{\code{codomain}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Specifies codomain of function.
Most importantly the tags of each output "Parameter" define whether it should
be minimized or maximized.  The default is to minimize each component.}
      \item{\code{check_values}}{(\code{logical(1)})\cr
Should points before the evaluation and the results be checked for validity?}
      \item{\code{rush}}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_points}{}}}
\subsection{\code{ArchiveAsync$push_points()}}{
  Push queued points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_points(xss, xss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_point}{}}}
\subsection{\code{ArchiveAsync$push_point()}}{
  Push a single queued point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_point(xs, xs_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_running_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_running_points}{}}}
\subsection{\code{ArchiveAsync$push_running_points()}}{
  Push running points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_running_points(xss, xss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_running_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_running_point}{}}}
\subsection{\code{ArchiveAsync$push_running_point()}}{
  Push running point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_running_point(xs, xs_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_finished_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_finished_points}{}}}
\subsection{\code{ArchiveAsync$push_finished_points()}}{
  Push finished points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_finished_points(xss, yss, xss_extra = NULL, yss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{yss}}{(list of named \code{list()})\cr
List of named lists of results.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
      \item{\code{yss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_finished_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_finished_point}{}}}
\subsection{\code{ArchiveAsync$push_finished_point()}}{
  Push a single finished point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_finished_point(xs, ys, xs_extra = NULL, ys_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{ys}}{(named \code{list()})\cr
Named list of results.}
      \item{\code{xs_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
      \item{\code{ys_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_failed_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_failed_points}{}}}
\subsection{\code{ArchiveAsync$push_failed_points()}}{
  Push failed points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_failed_points(xss, xss_extra = NULL, conditions = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
      \item{\code{conditions}}{(\code{list} | \code{NULL})\cr
List of conditions for each failed point.
If \code{NULL}, a generic error message is used.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_failed_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_failed_point}{}}}
\subsection{\code{ArchiveAsync$push_failed_point()}}{
  Push a single failed point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_failed_point(xs, xs_extra = NULL, condition = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
      \item{\code{condition}}{(\code{any} | \code{NULL})\cr
Condition of the failed point.
If \code{NULL}, a generic error message is used.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-pop_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-pop_point}{}}}
\subsection{\code{ArchiveAsync$pop_point()}}{
  Pop a point from the queue.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$pop_point()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-finish_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-finish_points}{}}}
\subsection{\code{ArchiveAsync$finish_points()}}{
  Save the results of multiple running points and move them to the finished points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$finish_points(keys, yss, x_domains, yss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character()})\cr
Keys of the points.}
      \item{\code{yss}}{(list of named \code{list()})\cr
List of named lists of results.}
      \item{\code{x_domains}}{(\code{list()})\cr
List of named lists of transformed point values.}
      \item{\code{yss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-finish_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-finish_point}{}}}
\subsection{\code{ArchiveAsync$finish_point()}}{
  Save the results of a running point and move it to the finished points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$finish_point(key, ys, x_domain, ys_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character(1)})\cr
Key of the point.}
      \item{\code{ys}}{(named \code{list()})\cr
Named list of results.}
      \item{\code{x_domain}}{(named \code{list()})\cr
Named list of transformed point values.}
      \item{\code{ys_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-fail_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-fail_points}{}}}
\subsection{\code{ArchiveAsync$fail_points()}}{
  Move multiple running points to the failed points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$fail_points(keys, conditions = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character()})\cr
Keys of the points.}
      \item{\code{conditions}}{(\code{list()} | \code{NULL})\cr
Conditions of the failed points.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-fail_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-fail_point}{}}}
\subsection{\code{ArchiveAsync$fail_point()}}{
  Move a running point to the failed points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$fail_point(key, condition = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character(1)})\cr
Key of the point.}
      \item{\code{condition}}{(\code{any} | \code{NULL})\cr
Condition of the failed point.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-push_result"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-push_result}{}}}
\subsection{\code{ArchiveAsync$push_result()}}{
  Deprecated.
Use \verb{$finish_point()} instead.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$push_result(key, ys, x_domain, extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character()})\cr
Key of the point.}
      \item{\code{ys}}{(\code{list()})\cr
Named list of results.}
      \item{\code{x_domain}}{(\code{list()})\cr
Named list of transformed point values.}
      \item{\code{extra}}{(\code{list()})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-data_with_state"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-data_with_state}{}}}
\subsection{\code{ArchiveAsync$data_with_state()}}{
  Fetch points with a specific state.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$data_with_state(
  fields = c("worker_id", "xs", "ys", "xs_extra", "ys_extra", "condition"),
  states = c("queued", "running", "finished", "failed")
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fields}}{(\code{character()})\cr
Fields to fetch.
Defaults to \code{c("worker_id", "xs", "ys", "xs_extra", "ys_extra", "condition")}.}
      \item{\code{states}}{(\code{character()})\cr
States of the tasks to be fetched.
Defaults to \code{c("queued", "running", "finished", "failed")}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-best"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-best}{}}}
\subsection{\code{ArchiveAsync$best()}}{
  Returns the best scoring evaluation(s).
For single-crit optimization, the solution that minimizes / maximizes the objective function.
For multi-crit optimization, the Pareto set / front.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$best(n_select = 1, ties_method = "first")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.
Ignored for multi-crit optimization.}
      \item{\code{ties_method}}{(\code{character(1L)})\cr
Method to break ties when multiple points have the same score.
Either \code{"first"} (default) or \code{"random"}.
Ignored for multi-crit optimization.
If \code{n_select > 1L}, the tie method is ignored and the first point is returned.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{\link[data.table:data.table]{data.table::data.table()}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-nds_selection"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-nds_selection}{}}}
\subsection{\code{ArchiveAsync$nds_selection()}}{
  Calculate best points w.r.t. non dominated sorting with hypervolume contribution.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$nds_selection(n_select = 1, ref_point = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n_select}}{(\code{integer(1L)})\cr
Amount of points to select.}
      \item{\code{ref_point}}{(\code{numeric()})\cr
Reference point for hypervolume.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{\link[data.table:data.table]{data.table::data.table()}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-clear}{}}}
\subsection{\code{ArchiveAsync$clear()}}{
  Clear all evaluation results from archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsync-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsync-clone}{}}}
\subsection{\code{ArchiveAsync$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsync$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
