% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveAsyncFrozen.R
\name{ArchiveAsyncFrozen}
\alias{ArchiveAsyncFrozen}
\title{Frozen Rush Data Storage}
\description{
Freezes the Redis data base of an \link{ArchiveAsync} to a  \code{data.table::data.table()}.
No further points can be added to the archive but the data can be accessed and analyzed.
Useful when the Redis data base is not permanently available.
Use the callback \link{bbotk.async_freeze_archive} to freeze the archive after the optimization has finished.
}
\section{S3 Methods}{

\itemize{
\item \code{as.data.table(archive)}\cr
\link{ArchiveAsync} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a tabular view of all performed function calls of the Objective.
The \code{x_domain} column is unnested to separate columns.
}
}

\examples{
# example only runs if a Redis server is available
if (mlr3misc::require_namespaces(c("rush", "redux", "mirai"), quietly = TRUE) &&
  redux::redis_available()) {
# define the objective function
fun = function(xs) {
  list(y = - (xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain
domain = ps(
  x1 = p_dbl(-10, 10),
  x2 = p_dbl(-5, 5)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# start workers
rush::rush_plan(worker_type = "mirai")
mirai::daemons(1)

# initialize instance
instance = oi_async(
  objective = objective,
  terminator = trm("evals", n_evals = 20),
  callback = clbk("bbotk.async_freeze_archive")
)

# load optimizer
optimizer = opt("async_random_search")

# trigger optimization
optimizer$optimize(instance)

# frozen archive
instance$archive

# best performing configuration
instance$archive$best()

# covert to data.table
as.data.table(instance$archive)
}
}
\seealso{
\link{ArchiveAsync}
}
\section{Super classes}{
\code{\link[bbotk:Archive]{Archive}} -> \code{\link[bbotk:ArchiveAsync]{ArchiveAsync}} -> \code{ArchiveAsyncFrozen}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all finished points.}

    \item{\code{queued_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all queued points.}

    \item{\code{running_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all running points.}

    \item{\code{finished_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all finished points.}

    \item{\code{failed_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Data table with all failed points.}

    \item{\code{n_queued}}{(\code{integer(1)})\cr
Number of queued points.}

    \item{\code{n_running}}{(\code{integer(1)})\cr
Number of running points.}

    \item{\code{n_finished}}{(\code{integer(1)})\cr
Number of finished points.}

    \item{\code{n_failed}}{(\code{integer(1)})\cr
Number of failed points.}

    \item{\code{n_evals}}{(\code{integer(1)})\cr
Number of evaluations stored in the archive.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ArchiveAsyncFrozen-initialize}{\code{ArchiveAsyncFrozen$new()}}
    \item \href{#method-ArchiveAsyncFrozen-push_points}{\code{ArchiveAsyncFrozen$push_points()}}
    \item \href{#method-ArchiveAsyncFrozen-push_point}{\code{ArchiveAsyncFrozen$push_point()}}
    \item \href{#method-ArchiveAsyncFrozen-push_running_points}{\code{ArchiveAsyncFrozen$push_running_points()}}
    \item \href{#method-ArchiveAsyncFrozen-push_running_point}{\code{ArchiveAsyncFrozen$push_running_point()}}
    \item \href{#method-ArchiveAsyncFrozen-push_finished_points}{\code{ArchiveAsyncFrozen$push_finished_points()}}
    \item \href{#method-ArchiveAsyncFrozen-push_finished_point}{\code{ArchiveAsyncFrozen$push_finished_point()}}
    \item \href{#method-ArchiveAsyncFrozen-push_failed_points}{\code{ArchiveAsyncFrozen$push_failed_points()}}
    \item \href{#method-ArchiveAsyncFrozen-push_failed_point}{\code{ArchiveAsyncFrozen$push_failed_point()}}
    \item \href{#method-ArchiveAsyncFrozen-pop_point}{\code{ArchiveAsyncFrozen$pop_point()}}
    \item \href{#method-ArchiveAsyncFrozen-finish_points}{\code{ArchiveAsyncFrozen$finish_points()}}
    \item \href{#method-ArchiveAsyncFrozen-finish_point}{\code{ArchiveAsyncFrozen$finish_point()}}
    \item \href{#method-ArchiveAsyncFrozen-fail_points}{\code{ArchiveAsyncFrozen$fail_points()}}
    \item \href{#method-ArchiveAsyncFrozen-fail_point}{\code{ArchiveAsyncFrozen$fail_point()}}
    \item \href{#method-ArchiveAsyncFrozen-push_result}{\code{ArchiveAsyncFrozen$push_result()}}
    \item \href{#method-ArchiveAsyncFrozen-data_with_state}{\code{ArchiveAsyncFrozen$data_with_state()}}
    \item \href{#method-ArchiveAsyncFrozen-clear}{\code{ArchiveAsyncFrozen$clear()}}
    \item \href{#method-ArchiveAsyncFrozen-clone}{\code{ArchiveAsyncFrozen$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="format"><a href='../../bbotk/html/Archive.html#method-Archive-format'><code>Archive$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="help"><a href='../../bbotk/html/Archive.html#method-Archive-help'><code>Archive$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="print"><a href='../../bbotk/html/Archive.html#method-Archive-print'><code>Archive$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="ArchiveAsync" data-id="best"><a href='../../bbotk/html/ArchiveAsync.html#method-ArchiveAsync-best'><code>ArchiveAsync$best()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="bbotk" data-topic="ArchiveAsync" data-id="nds_selection"><a href='../../bbotk/html/ArchiveAsync.html#method-ArchiveAsync-nds_selection'><code>ArchiveAsync$nds_selection()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-initialize}{}}}
\subsection{\code{ArchiveAsyncFrozen$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$new(archive)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{archive}}{(\link{ArchiveAsync})\cr
The archive to freeze.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_points()}}{
  Push queued points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_points(xss, xss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_point()}}{
  Push a single queued point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_point(xs, xs_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_running_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_running_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_running_points()}}{
  Push running points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_running_points(xss, xss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_running_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_running_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_running_point()}}{
  Push running point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_running_point(xs, xs_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_finished_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_finished_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_finished_points()}}{
  Push finished points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_finished_points(
  xss,
  yss,
  xss_extra = NULL,
  yss_extra = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{yss}}{(list of named \code{list()})\cr
List of named lists of results.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
      \item{\code{yss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_finished_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_finished_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_finished_point()}}{
  Push a single finished point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_finished_point(
  xs,
  ys,
  xs_extra = NULL,
  ys_extra = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{ys}}{(named \code{list()})\cr
Named list of results.}
      \item{\code{xs_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
      \item{\code{ys_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_failed_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_failed_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_failed_points()}}{
  Push failed points to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_failed_points(xss, xss_extra = NULL, conditions = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xss}}{(list of named \code{list()})\cr
List of named lists of point values.}
      \item{\code{xss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
      \item{\code{conditions}}{(\code{list} | \code{NULL})\cr
List of conditions for each failed point.
If \code{NULL}, a generic error message is used.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_failed_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_failed_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_failed_point()}}{
  Push a single failed point to the archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_failed_point(xs, xs_extra = NULL, condition = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{xs}}{(named \code{list()})\cr
Named list of point values.}
      \item{\code{xs_extra}}{(\verb{named list()} | \code{NULL})\cr
Named list of additional information.}
      \item{\code{condition}}{(\code{any} | \code{NULL})\cr
Condition of the failed point.
If \code{NULL}, a generic error message is used.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-pop_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-pop_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$pop_point()}}{
  Pop a point from the queue.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$pop_point()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-finish_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-finish_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$finish_points()}}{
  Save the results of multiple running points and move them to the finished points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$finish_points(keys, yss, x_domains, yss_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character()})\cr
Keys of the points.}
      \item{\code{yss}}{(list of named \code{list()})\cr
List of named lists of results.}
      \item{\code{x_domains}}{(\code{list()})\cr
List of named lists of transformed point values.}
      \item{\code{yss_extra}}{(list of named \code{list()} | \code{NULL})\cr
List of named lists of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-finish_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-finish_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$finish_point()}}{
  Save the results of a running point and move it to the finished points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$finish_point(key, ys, x_domain, ys_extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character(1)})\cr
Key of the point.}
      \item{\code{ys}}{(named \code{list()})\cr
Named list of results.}
      \item{\code{x_domain}}{(named \code{list()})\cr
Named list of transformed point values.}
      \item{\code{ys_extra}}{(named \code{list()} | \code{NULL})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-fail_points"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-fail_points}{}}}
\subsection{\code{ArchiveAsyncFrozen$fail_points()}}{
  Move multiple running points to the failed points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$fail_points(keys, conditions = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{keys}}{(\code{character()})\cr
Keys of the points.}
      \item{\code{conditions}}{(\code{list()} | \code{NULL})\cr
Conditions of the failed points.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-fail_point"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-fail_point}{}}}
\subsection{\code{ArchiveAsyncFrozen$fail_point()}}{
  Move a running point to the failed points.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$fail_point(key, condition = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character(1)})\cr
Key of the point.}
      \item{\code{condition}}{(\code{any} | \code{NULL})\cr
Condition of the failed point.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-push_result"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-push_result}{}}}
\subsection{\code{ArchiveAsyncFrozen$push_result()}}{
  Deprecated.
Use \verb{$finish_point()} instead.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$push_result(key, ys, x_domain, extra = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{key}}{(\code{character()})\cr
Key of the point.}
      \item{\code{ys}}{(\code{list()})\cr
Named list of results.}
      \item{\code{x_domain}}{(\code{list()})\cr
Named list of transformed point values.}
      \item{\code{extra}}{(\code{list()})\cr
Named list of additional information.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-data_with_state"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-data_with_state}{}}}
\subsection{\code{ArchiveAsyncFrozen$data_with_state()}}{
  Fetch points with a specific state.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$data_with_state(
  fields = c("xs", "ys", "xs_extra", "worker_extra", "ys_extra", "condition"),
  states = c("queued", "running", "finished", "failed"),
  reset_cache = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fields}}{(\code{character()})\cr
Fields to fetch.
Defaults to \code{c("xs", "ys", "xs_extra", "worker_extra", "ys_extra")}.}
      \item{\code{states}}{(\code{character()})\cr
States of the tasks to be fetched.
Defaults to \code{c("queued", "running", "finished", "failed")}.}
      \item{\code{reset_cache}}{(\code{logical(1)})\cr
Whether to reset the cache of the finished points.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-clear"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-clear}{}}}
\subsection{\code{ArchiveAsyncFrozen$clear()}}{
  Clear all evaluation results from archive.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$clear()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArchiveAsyncFrozen-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArchiveAsyncFrozen-clone}{}}}
\subsection{\code{ArchiveAsyncFrozen$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ArchiveAsyncFrozen$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
