% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{assign_result_default}
\alias{assign_result_default}
\title{Default assign_result function}
\usage{
assign_result_default(inst)
}
\arguments{
\item{inst}{\link{OptimInstance}}
}
\description{
Default assign_result function
}
\keyword{internal}
