% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{assign_result_default}
\alias{assign_result_default}
\alias{assign_result_default.OptimInstance}
\alias{assign_result_default.OptimInstanceRush}
\title{Default Assign Result Function}
\usage{
assign_result_default(inst)

\method{assign_result_default}{OptimInstance}(inst)

\method{assign_result_default}{OptimInstanceRush}(inst)
}
\arguments{
\item{inst}{\link{OptimInstance}}
}
\description{
Used internally in the \link{Optimizer}.
It is the default way to determine the result by simply obtaining the best performing result from the archive.
}
\keyword{internal}
