% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{bbotk-package}
\alias{bbotk}
\alias{bbotk-package}
\title{bbotk: Black-Box Optimization Toolkit}
\description{
Implements toolkit functions for black-box
    optimization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/bbotk}
  \item Report bugs at \url{https://github.com/mlr-org/bbotk/issues}
}

}
\author{
\strong{Maintainer}: Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Martin Binder \email{martin.binder@mail.com}
  \item Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})
}

}
