% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bbotk_worker_loop}
\alias{bbotk_worker_loop}
\title{Worker loop for Rush}
\usage{
bbotk_worker_loop(rush, objective, search_space)
}
\arguments{
\item{rush}{\link[rush:Rush]{rush::Rush}\cr
Rush.}

\item{objective}{\link{Objective}\cr
Objective function.}

\item{search_space}{\link[paradox:ParamSet]{paradox::ParamSet}\cr
Search space.}
}
\description{
Loop run on the workers.
Pops a task from the queue and evaluates it with the objective function.
Pushes the results back to the data base.
}
