% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_loops.R
\name{bbotk_worker_loop}
\alias{bbotk_worker_loop}
\title{Worker loop for Rush}
\usage{
bbotk_worker_loop(rush, optimizer, instance)
}
\arguments{
\item{rush}{(\code{Rush})\cr
If a rush instance is supplied, the tuning runs without batches.}

\item{optimizer}{\link{OptimizerAsync}.}

\item{instance}{\link{OptimInstanceAsync}.}
}
\description{
Loop run on the workers.
Pops a task from the queue and evaluates it with the objective function.
Pushes the results back to the data base.
}
