% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{branin}
\alias{branin}
\title{Branin Function}
\source{
Wu J, Toscano-Palmerin S, Frazier PI, Wilson AG (2019).
\dQuote{Practical Multi-fidelity Bayesian Optimization for Hyperparameter Tuning.}
1903.04703.
}
\usage{
branin(xs)
}
\arguments{
\item{xs}{List with the input for a single point
(e.g. \code{list(x1 = 1, x2 = 2, fidelity = 0.5)}).}
}
\value{
\code{list(1)}
}
\description{
Augmented 2-D Branin function with fidelity parameter.
}
\examples{
branin(list(x1 = 12, x2 = 2, fidelity = 1))
}
