% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_dominated}
\alias{is_dominated}
\title{Calculate which points are dominated}
\usage{
is_dominated(ymat)
}
\arguments{
\item{ymat}{(\code{matrix()}) \cr
A numeric matrix. Each column (!) contains one point.}
}
\description{
Calculates which points are not dominated,
i.e. points that belong to the Pareto front.
}
