% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{local_search}
\alias{local_search}
\title{Local Search}
\usage{
local_search(objective, search_space, control = local_search_control())
}
\arguments{
\item{objective}{\link{function(dt)}
Objective to optimize.}

\item{search_space}{\link{ParamSet}
Search space for decision variables.}
}
\value{
\link{data.table}
}
\description{
Local search optimizer
}
