% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{local_search_control}
\alias{local_search_control}
\title{Local Search Control}
\usage{
local_search_control(
  minimize = TRUE,
  n_searches = 10L,
  n_steps = 5L,
  n_neighs = 10L,
  mut_sd = 0.1
)
}
\arguments{
\item{n_searches}{\link{integer(1)}
Number of local searches.}

\item{n_steps}{\link{integer(1)}
Number of steps per local search.}

\item{n_neighs}{\link{integer(1)}
Number of neighbors per local search.}

\item{mut_sd}{\link{numeric(1)}
Standard deviation of the mutation.}
}
\value{
\link{local_search_control}
List with control params as S3 object.
}
\description{
Control parameters for local search optimizer
}
