% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatchSmac3.R
\name{mlr_optimizers_smac}
\alias{mlr_optimizers_smac}
\alias{OptimizerBatchSmac3}
\title{Sequential Model-Based Algorithm Configuration (SMAC3)}
\source{
Lindauer M, Eggensperger K, Feurer M, Biedenkapp A, Deng D, Benjamins C, Ruhkopf T, Sass R, Hutter F (2022).
\dQuote{SMAC3: A Versatile Bayesian Optimization Package for Hyperparameter Optimization.}
\emph{Journal of Machine Learning Research}, \bold{23}(54), 1--9.
\url{http://jmlr.org/papers/v23/21-0888.html}.
}
\description{
Calls SMAC3 from Python via the \CRANpkg{reticulate} package.
}
\note{
All parameters of the search space must have default values.
}
\section{Parameters}{

\describe{
\item{\code{n_init}}{\code{integer(1)}\cr
Number of initial configurations to evaluate before starting the optimization.
Defaults to \code{10} times the number of hyperparameters.}
\item{\code{facade}}{\code{character(1)}\cr
Facade to use.
Either \code{"smac4bb"} (Black-Box Facade) or \code{"smac4hb"} (Hyperparameter Optimization Facade).
Default is \code{"smac4bb"}.}
\item{\code{output_directory}}{\code{character(1)}\cr
Directory to store the output of SMAC3.
Default is a temporary directory.}
}
}

\section{Progress Bars}{

\verb{$optimize()} supports progress bars via the package \CRANpkg{progressr}
combined with a \link{Terminator}. Simply wrap the function in
\code{progressr::with_progress()} to enable them. We recommend to use package
\CRANpkg{progress} as backend; enable with \code{progressr::handlers("progress")}.
}

\examples{
\dontrun{
# define the objective function
fun = function(xs) {
  list(y = -(xs[[1]] - 2)^2 - (xs[[2]] + 3)^2 + 10)
}

# set domain (all parameters must have defaults for ConfigSpace)
domain = ps(
  x1 = p_dbl(-10, 10, default = 0),
  x2 = p_dbl(-5, 5, default = 0)
)

# set codomain
codomain = ps(
  y = p_dbl(tags = "maximize")
)

# create objective
objective = ObjectiveRFun$new(
  fun = fun,
  domain = domain,
  codomain = codomain,
  properties = "deterministic"
)

# initialize instance
instance = oi(
  objective = objective,
  terminator = trm("evals", n_evals = 20)
)

# load optimizer
optimizer = opt("smac")

# trigger optimization
optimizer$optimize(instance)

# all evaluated configurations
instance$archive

# best performing configuration
instance$result
}
}
\section{Super classes}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{\link[bbotk:OptimizerBatch]{bbotk::OptimizerBatch}} -> \code{OptimizerBatchSmac3}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerBatchSmac3-new}{\code{OptimizerBatchSmac3$new()}}
\item \href{#method-OptimizerBatchSmac3-clone}{\code{OptimizerBatchSmac3$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="OptimizerBatch" data-id="optimize"><a href='../../bbotk/html/OptimizerBatch.html#method-OptimizerBatch-optimize'><code>bbotk::OptimizerBatch$optimize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchSmac3-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchSmac3-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchSmac3$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerBatchSmac3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerBatchSmac3-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerBatchSmac3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
