% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorCombo.R
\name{mlr_terminators_combo}
\alias{mlr_terminators_combo}
\alias{TerminatorCombo}
\title{Combine Terminators}
\arguments{
\item{terminators}{(\code{list()})\cr
List of objects of class \link{Terminator}.}

\item{instance}{(\link{TuningInstance}).}
}
\value{
\code{logical(1)}.
}
\description{
This class takes multiple \link{Terminator}s and terminates as soon as one or all
of the included terminators are positive.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE} otherwise.
}
\section{Fields}{

\describe{
\item{\code{terminators}}{(\code{list()})\cr
List of objects of class \link{Terminator}.}
}}

\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("combo")
term("combo")
}
}

\section{Parameters}{

\itemize{
\item \code{any} (\code{logical(1)})\cr
Terminate iff any included terminator is positive? (not all), default is `TRUE.
}
}

\examples{
term("combo",
  list(term("model_time", secs = 60), term("evals", n_evals = 10)),
  any = FALSE
)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_model_time}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
