% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorRunTime.R
\name{mlr_terminators_run_time}
\alias{mlr_terminators_run_time}
\alias{TerminatorRunTime}
\title{Terminator that stops according to the run time}
\description{
Class to terminate the optimization after the optimization process took a
number of seconds on the clock.
}
\section{Dictionary}{

This \link{Terminator} can be instantiated via the
\link[mlr3misc:Dictionary]{dictionary} \link{mlr_terminators} or with the associated
sugar function \code{\link[=term]{term()}}:\preformatted{mlr_terminators$get("run_time")
term("run_time")
}
}

\section{Parameters}{

\itemize{
\item \code{secs} \code{numeric(1)}\cr
Maximum allowed time, in seconds, default is 30.
}
}

\examples{
term("run_time", secs = 1800)
}
\seealso{
Other Terminator: 
\code{\link{Terminator}},
\code{\link{mlr_terminators_clock_time}},
\code{\link{mlr_terminators_combo}},
\code{\link{mlr_terminators_evals}},
\code{\link{mlr_terminators_none}},
\code{\link{mlr_terminators_perf_reached}},
\code{\link{mlr_terminators_stagnation_batch}},
\code{\link{mlr_terminators_stagnation}},
\code{\link{mlr_terminators}}
}
\concept{Terminator}
\section{Super class}{
\code{\link[bbotk:Terminator]{bbotk::Terminator}} -> \code{TerminatorRunTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorRunTime$new()}}
\item \href{#method-is_terminated}{\code{TerminatorRunTime$is_terminated()}}
\item \href{#method-progressr_steps}{\code{TerminatorRunTime$progressr_steps()}}
\item \href{#method-progressr_update}{\code{TerminatorRunTime$progressr_update()}}
\item \href{#method-clone}{\code{TerminatorRunTime$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Terminator" data-id="format">}\href{../../bbotk/html/Terminator.html#method-format}{\code{bbotk::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Terminator" data-id="print">}\href{../../bbotk/html/Terminator.html#method-print}{\code{bbotk::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorRunTime$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\if{latex}{\out{\hypertarget{method-is_terminated}{}}}
\subsection{Method \code{is_terminated()}}{
Is \code{TRUE} iff the termination criterion is positive, and \code{FALSE}
otherwise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorRunTime$is_terminated(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{(\link{Archive}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical(1)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progressr_steps"></a>}}
\if{latex}{\out{\hypertarget{method-progressr_steps}{}}}
\subsection{Method \code{progressr_steps()}}{
Returns the allowed runtime in seconds.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorRunTime$progressr_steps(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{(\link{Archive}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{integer(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progressr_update"></a>}}
\if{latex}{\out{\hypertarget{method-progressr_update}{}}}
\subsection{Method \code{progressr_update()}}{
Returns time difference between the last two batches (\code{amount}) and
current runtime (\code{sum}) in seconds.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorRunTime$progressr_update(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{(\link{Archive}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of \code{numeric(1)} and \code{integer(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorRunTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
