% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{Objective}
\alias{Objective}
\title{Map}
\description{
Describes a function (domain) -> (codomain)

domain, codomain: ParamSet
fun: function. takes 1 list argument and returns a list, with positions and names agreeing with
domain, codomain

fun must be given as a function of two arguments. The second is set from
the param_set. $fun will then only have one parameter.
}
