% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{Objective}
\alias{Objective}
\title{Objective function with domain and co-domain}
\arguments{
\item{id}{\code{character(1)}}

\item{properties}{\code{character()}}

\item{domain}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{codomain}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{xs}{\code{list()} A list that contains a single x value, e.g. \code{list(x1 = 1, x2 = 2)}.}

\item{xss}{\code{list()} A list of lists that contains multiple x values, e.g. \code{list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4))}.}
}
\value{
\code{character()}

\code{character()}

\code{list()} A list that contains the result of the evaluation, e.g. \code{list(y = 1)}.

\code{list()} A list that contains the result of all evaluations, e.g. \code{list(list(y = 1), list(y=2))}.

\code{list()} A list that contains the result of the evaluation, e.g. \code{list(y = 1)}.

\code{list()} A list that contains the result of all evaluations, e.g. \code{list(list(y = 1), list(y=2))}.
}
\description{
Describes a black-box objective function that maps an arbitrary domain to a numerical codomain.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Helper for print outputs.

Print method.

Evaluates a single input value on the objective function

Evaluates multiple input values on the objective function

Evaluates a single input value on the objective function and checks its validity as well as the validity of the result.

Evaluates multiple input values on the objective function and checks the validity of the input.
}
\section{Fields}{

\describe{
\item{\code{id}}{\code{character(1)}}

\item{\code{properties}}{\code{character()}}

\item{\code{domain}}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{\code{codomain}}{\link[paradox:ParamSet]{paradox::ParamSet}}
}}

