% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimInstanceMulticrit.R
\name{OptimInstanceMulticrit}
\alias{OptimInstanceMulticrit}
\title{Optimization Instance with budget and archive}
\description{
Wraps an multi-criteria \link{Objective} function with extra services for convenient evaluation.
Inherits from \link{OptimInstance}
}
\section{Super class}{
\code{\link[bbotk:OptimInstance]{bbotk::OptimInstance}} -> \code{OptimInstanceMulticrit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result_opt_x}}{\code{list()}\cr
(transformed) x part of the result in the \emph{domain space} of the objective.}

\item{\code{result_y}}{\code{numeric(1)}
Optimal outcome.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-assign_result}{\code{OptimInstanceMulticrit$assign_result()}}
\item \href{#method-clone}{\code{OptimInstanceMulticrit$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="eval_batch">}\href{../../bbotk/html/OptimInstance.html#method-eval_batch}{\code{bbotk::OptimInstance$eval_batch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="format">}\href{../../bbotk/html/OptimInstance.html#method-format}{\code{bbotk::OptimInstance$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="initialize">}\href{../../bbotk/html/OptimInstance.html#method-initialize}{\code{bbotk::OptimInstance$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="OptimInstance" data-id="print">}\href{../../bbotk/html/OptimInstance.html#method-print}{\code{bbotk::OptimInstance$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_result"></a>}}
\if{latex}{\out{\hypertarget{method-assign_result}{}}}
\subsection{Method \code{assign_result()}}{
The \link{Optimizer} object writes the best found points
and estimated performance values here (e.g. the Pareto Front). For internal use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceMulticrit$assign_result(xdt, ydt, opt_x = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{\link[data.table:data.table]{data.table::data.table} :: set of untransformed points / points from the \emph{search space} that lead to the result of the optimization}

\item{\code{ydt}}{\link[data.table:data.table]{data.table::data.table} :: Optimal outcome.}

\item{\code{opt_x}}{\code{list()} :: Transformed x values / points from the \emph{domain} of the \link{Objective} as a named list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimInstanceMulticrit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
