% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsync.R
\name{optimize_async_default}
\alias{optimize_async_default}
\title{Default Asynchronous Optimization}
\usage{
optimize_async_default(instance, optimizer, design = NULL)
}
\arguments{
\item{instance}{\link{OptimInstanceAsync}.}

\item{optimizer}{\link{OptimizerAsync}.}
}
\description{
Used internally in \link{OptimizerAsync}.
}
\keyword{internal}
