% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerBatch.R
\name{optimize_batch_default}
\alias{optimize_batch_default}
\title{Default Batch Optimization Function}
\usage{
optimize_batch_default(inst, self, private)
}
\arguments{
\item{inst}{\link{OptimInstance}}

\item{self}{\link{OptimizerBatch}}

\item{private}{(\code{environment()})}
}
\value{
\link[data.table:data.table]{data.table::data.table}
}
\description{
Used internally in the \link{OptimizerBatch}.
Brings together the private \code{.optimize()} method and the private \code{.assign_result()} method.
}
\keyword{internal}
