% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{Optimizer}
\alias{Optimizer}
\title{Optimizer}
\arguments{
\item{param_set}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{param_classes}{\code{character()}}

\item{properties}{\code{character()}}

\item{packages}{\code{character()}}

\item{inst}{\link{OptimInstance}}
}
\value{
\code{character()}
}
\description{
Abstract \code{Optimizer} class that implements the base functionality each
\code{Optimizer} subclass must provide. A \code{Optimizer} object describes the
optimization strategy.

A \code{Optimizer} object must write its result to the \verb{$assign_result()} method
of the \link{OptimInstance} at the end in order to store the best point  and its
estimated performance vector.

Creates a new instance of this \link[R6:R6Class]{R6} class.

Helper for print outputs.

Print method.

Performes the optimization.
}
\section{Fields}{

\describe{
\item{\code{param_set}}{\link[paradox:ParamSet]{paradox::ParamSet}}

\item{\code{param_classes}}{\code{character()}}

\item{\code{properties}}{\code{character()}}

\item{\code{packages}}{\code{character()}}
}}

